Clean up error checking and handling.
The TLV length includes the T and the V, so it must be at least 4; fail
if it's not.

If the IPv4 or IPv6 prefix length isn't valid, don't bother trying to
parse the rest of the addresses, just give up.

Don't bother returning offset values that aren't used.

Rename some routines to indicate that they process more than one item.

Add some comments while we're at it.

Change-Id: I6825216f0e7218e230d8d60d958b3c2453a3bb62
Reviewed-on: https://code.wireshark.org/review/13016
Reviewed-by: Guy Harris <guy@alum.mit.edu>