Qt: Update some keyboard shortcuts to match the GTK+ UI.
As the QKeySequence documentation says,

  "On Mac OS X, references to "Ctrl", Qt::CTRL, Qt::Control and
   Qt::ControlModifier correspond to the Command keys on the Macintosh
   keyboard, and references to "Meta", Qt::META, Qt::Meta and
   Qt::MetaModifier correspond to the Control keys. Developers on Mac OS
   X can use the same shortcut descriptions across all platforms, and
   their applications will automatically work as expected on Mac OS X."

This also applies to Qt Creator on OS X. If you assign a shortcut to an
action that contains the Control key, it will draw the ^ symbol in the
UI but will save "Meta" in the .ui file instead of "Ctrl", in the manner
of a well-meaning-but-not-helpful comedy sidekick.

This happened for the actions listed below. Replace "Meta" in their
shortcuts with "Ctrl".

- Unmark all (Ctrl+Alt+M)
- Next marked packet (Ctrl+Shift+N)
- Previous marked packet (Ctrl+Shift+B)
- Show packet times... (Ctrl+Alt+1 - Ctrl+Alt+8)

This matches the GTK+ UI on Windows and Linux, and uses the Command key
on OS X. If we really want to use the Control key everywhere we can
override the action sequences in main_window.cpp. We might want to do
this for the "mark" actions since Command+M is the standard key for
"Minimize this window".

Change-Id: I1537cee5bc27a32b505bace01c1de3703a18dd6a
Reviewed-on: https://code.wireshark.org/review/13238
Reviewed-by: Gerald Combs <gerald@wireshark.org>
Petri-Dish: Gerald Combs <gerald@wireshark.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
(cherry picked from commit 1f0c9f669375da29c60552ce2f7a63c9bfbd1433)
Reviewed-on: https://code.wireshark.org/review/13252