Assorted cleanups.
Fix indentation.

Just directly assign values to elements in the packet buffer; no need to
convert them to numbers and note the value as a comment.

Give more detail in the comment for null-terminating buffers.  Terminate
packet_buf[] once we're finished reading into it, to make it a bit
clearer what's being done.

Make the magic number buffer 513 bytes, so we have 512 bytes plus a
terminating null.

Change-Id: Ie182d93393cc55835b24075e908393c386c85c24
Reviewed-on: https://code.wireshark.org/review/13250
Reviewed-by: Guy Harris <guy@alum.mit.edu>