Handle frames where the type field is a length field.
The BNEP Version 1.0 spec says, for BNEP_GENERAL_ETHERNET packets,
"Note: Networking Protocol Types as used in this specification SHALL be
taken to include values in the range 0x0000-0x05dc, used to represent
the IEEE802.3 length interpretation of the IEEE802.3 length/type
field.", although it says that it's not mandatory to process those
packets.

We'll process them.

We don't need an Ethernet dissector handle, so don't fetch one.

Change-Id: I10b1fce62688fc0643435d0a01bdb969577df97e
Reviewed-on: https://code.wireshark.org/review/13376
Reviewed-by: Guy Harris <guy@alum.mit.edu>