Separate "Ethernet in capture file" and "Ethernet maybe with FCS" dissectors.
Have a dissector that is passed a "struct eth_phdr" pointer, indicating
whether there is an FCS, there is no FCS, or there's maybe an FCS, and
an "eth_maybefcs" dissector, to be called from other dissectors.  The
latter takes no data argument.

That obviates the need for callers of the latter to fill in an
"eth_phdr" structure.

Note in a comment that setting the "assume an FCS" preference overrides
a file format handler in Wiretap saying "we have no FCS".  I seem to
remember that this might be intentional.

Change-Id: Ied48a3c0ae2b2ff1010032aa5a4a97fe394d48f6
Ping-Bug: 9933
Reviewed-on: https://code.wireshark.org/review/15388
Reviewed-by: Guy Harris <guy@alum.mit.edu>