Separate "Ethernet in capture file" and "Ethernet maybe with FCS" dissectors.
Have a dissector that is passed a "struct eth_phdr" pointer, indicating
whether there is an FCS, there is no FCS, or there's maybe an FCS, and
an "eth_maybefcs" dissector, to be called from other dissectors.  The
latter takes no data argument.

That obviates the need for callers of the latter to fill in an
"eth_phdr" structure.

Note in a comment that setting the "assume an FCS" preference overrides
a file format handler in Wiretap saying "we have no FCS".  I seem to
remember that this might be intentional.

Ping-Bug: 9933
Change-Id: I600e1351d468ab31d48369edb96832d6da3e480c
Reviewed-on: https://code.wireshark.org/review/13432
Reviewed-by: Guy Harris <guy@alum.mit.edu>