Fix handling of the magic number.
Just treat it as an array of bytes.  When checking for whether it's a
pcap file, also determine whether it's big-endian or little-endian.
Note that reading it in *host* byte order will tell you whether it's in
your byte order or byte-swapped; you have to know your byte order to
know whether that means little-endian or big-endian.

Have a #define for the magic number size, as all magic number values
must be that size, and use that as the size of the magic-number arrays.

Handle nanosecond timestamp resolution while we're at it.

Change-Id: I4d83579d919ae9f15888afca14317631d413ca51
Reviewed-on: https://code.wireshark.org/review/13490
Reviewed-by: Guy Harris <guy@alum.mit.edu>