Fix handling of the byte order magic number.
Just treat it as an array of bytes.  When checking for whether it's a
pcapng file, also determine whether it's big-endian or little-endian.
Note that reading it in *host* byte order will tell you whether it's in
your byte order or byte-swapped; you have to know your byte order to
know whether that means little-endian or big-endian.

Have a #define for the byte-order magic number size, as all byte order
magic number values must be that size, and use that as the size of the
magic-number arrays.

Also use a #define for the SHB block type magic number.

Get rid of a now-unused expert info.  (If the magic number isn't
something we recognize, we don't treat the file as a pcap file, so it
can never be "unknown".)

Change-Id: Ic74cceac17d1490eb70a28f67cb4dbb512e031ac
Reviewed-on: https://code.wireshark.org/review/13494
Reviewed-by: Guy Harris <guy@alum.mit.edu>