Check whether Ness is valid as well.
It's also used as an array index, and it also is no guaranteed to be in
the proper range in all capture files, so bounds-check it.

Make the variable's type match that of the field in the radio metadata.
(It's never negative, so it's unsigned.)

This appears to be the last of the unchecked array references from bug
12085.

Bug: 12085
Change-Id: I3e5b821a9b70dfd520d01036f6f7d02008aca200
Reviewed-on: https://code.wireshark.org/review/13806
Reviewed-by: Guy Harris <guy@alum.mit.edu>