Dissect the frame data.
We just dissect it as raw bytes for now; ultimately, we need to process
it the same way we process data for other forms of USB capture.

This also catches the case where the frame length is bogusly large
(including so large that rounding it up to a multiple of 4 overflows).

Bug: 12153
Change-Id: I537974d548fdcda917d9fce8189eb2134bc17bb9
Reviewed-on: https://code.wireshark.org/review/14103
Reviewed-by: Guy Harris <guy@alum.mit.edu>