Dissect the FCS after dissecting the header.
First dissect the address field, then dissect the control field, then
dissect the FCS if present, then dissect the info field.

That makes the dissection more like the dissection of other protocols
with an FCS at the end, and means that we don't throw an exception too
early.

Remove the "MLT CHANGES" comments - if somebody cares who did what, they
can browse the commit history; it's not as if the version prior to those
changes was somehow the "pure" version and that we need to keep the
changes carefully demarcated.

Get rid of no-longer-necessary variable.

Change-Id: I249440971e64ecbb0959ebbea1b2897a2e12375a
Reviewed-on: https://code.wireshark.org/review/14293
Reviewed-by: Guy Harris <guy@alum.mit.edu>